requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
packaging==24.1
pillow==10.3.0
pluggy==1.5.0
prompt_toolkit==3.0.47
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
referencing==0.35.1
rpds-py==0.18.1
setuptools==70.1.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
